+++ src/lib/interfaces/papa-parse-global-config.ts
@@ -1,0 +1,3 @@
+export interface PapaParseGlobalConfig {
+    scriptPath?: string;
+}